Update Main.py
Line 365- There was a check missing for the customer payment if they paid the equivalent. The solution is that I add the condition for it to be included. The try and handling are still acting if it's a negative float or and char or symbols that do not equate to paying more or its the exact amount.
Line 208-216- The issue with the main dish was that it threw an error for any non-integer or symbol. Line 208 had its text modified but Line 216 is where I added another try and catch for specifically the chars, and symbols that aren't integers. Currently the code the will now handle the input.